Broke out some code that was common across multiple views for additional info, company logo and quit buttons
Making a start on cleaning up SimpleMode layout
